backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.3",
    "polyfactory>=3.3.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.37",
    "supabase>=2.28.3",
    "uvicorn>=0.42.0",
]
